tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/bufferedAmount/bufferedAmount-unicode.html

<!doctype html>
<meta charset=utf-8>
<title>WebSockets: bufferedAmount for unicode data</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  var data = "¥¥¥¥¥¥";
  ws.onopen = t.step_func(function(e) {
    ws.send(data);
    assert_equals(data.length * 2, ws.bufferedAmount);
  });
  ws.onmessage = t.step_func(function(e) {
    assert_equals(e.data, data);
    t.done();
  });
});
</script>
