LayoutTests/platform/mac-bigsur-wk2/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/bufferedAmount/bufferedAmount-unicode-expected.txt


FAIL WebSockets: bufferedAmount for unicode data assert_equals: expected 0 but got 12
